requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
passlib==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
ruff==0.1.6
black==23.11.0
isort==5.12.0
pytest==7.4.3
pytest-flask==1.3.0
